requirements.txt
django
django-admin
djangorestframework
django-cors-headers